courses/udemy_complete_genai/3.chatbots/3.streamlit/requirements.txt
streamlit
python-dotenv
langchain-community
langchain-ollama
